requirements.txt
boto3
spacy
fastapi
python-dotenv
psycopg2-binary
pydantic
pandas
openpyxl
reportlab
